Started creating an application Bootstrapper
Each module can have an IBootstrapperTask

In your main application you will have to reference all
necessary assemblies as string and pass it to Bootstraper.
ExecuteBoostrapperTasks(<assemblyName>)
It will then search for all IBootstrapperTasks in that
assembly and pass the IContainerBuilder forward so that
the assembly can register all necessary types.