IBootstrapperTask Assemblies can now be registered
Now each assembly can define for itself which interfaces
will be exported to DI and can be used in the program.
The IBootstrapperTask can also be used to configure modules
and plug stuff together.
However if I register the IUserService in F# there seems to be
a problem. If I register the same type in C# it works.
Have to investigate on that a little bit more :)